キャラクターをフィルタする関数を別モジュールに切り分け
